% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bootstrap.parameters.R
\name{bootstrap.parameters}
\alias{bootstrap.parameters}
\title{Title}
\usage{
bootstrap.parameters(x, nboot = 1000, upper = NULL, lower = NULL)
}
\arguments{
\item{nboot}{}

\item{fit}{}
}
\description{
Title
}
\examples{
NULL
}

