% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_parameters_element}
\alias{bootstrap_parameters_element}
\title{Create a bootstrap distribution of a particular coefficient from a model fit for elementfit}
\usage{
bootstrap_parameters_element(x,nboot,upper,lower,...)
}
\arguments{
\item{x}{an object of class "\code{elementfit}"}

\item{nboot}{number of bootstrap replications}

\item{upper}{optional vector of upper bounds for the bootstrap replicates}

\item{lower}{optional vector of lower bounds for the bootstrap replicates}

\item{...}{passed to \code{\link{optim}}}
}
\value{
returns a \code{litfit_bootstrap object}
}
\description{
Create a bootstrap distribution of a particular coefficient from a model fit for elementfit
}
\examples{
fit <- fit_element(mass.remaining=pineneedles$Mass.remaining,
element.remaining=c(1.1,1.2,1.3,1,0.8,0.6),
model="import.model",iters=1000)
boot1 <- bootstrap_parameters_element(fit, nboot = 500)



}
