% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim-fxns.R
\name{fit_element}
\alias{fit_element}
\title{Fit different models to single cohort element data}
\usage{
fit_element(
  mass.remaining,
  element.remaining,
  model = c("import.model"),
  iters = 500,
  upper = NULL,
  lower = NULL,
  ...
)
}
\arguments{
\item{mass.remaining}{proportional mass loss, that is, mi/m0}

\item{element.remaining}{proportional element remaining ni/n0}

\item{model}{there is one model currently implemented (see below)}

\item{iters}{Number of random starts for the fitting.  Use higher numbers for models with larger numbers of parameters and for models that inherently tend to be less identifiable.}

\item{upper, lower}{\bold{Optional} user specified values for the upper and lower bounds used by
\code{optim} in the fitting procedure. Use with care, only minimal sanity checking is currently
 implemented.}

\item{...}{Additional arguments passed to \code{\link{optim}}}
}
\value{
returns a elementfit object with the following elements: 
 \itemize{
\item{\bold{optimFit:} a list generated by the \code{\link{optim}} function}
\item{\bold{logLik:} the log-likelihood of the model}
\item{\bold{time:} vector of time (same as input)}
\item{\bold{mass:} vector os mass remaining (same as input)}
\item{\bold{predicted:} predicted values from the model for each of the points within time}
\item{\bold{model:} name of the model }
\item{\bold{nparams:} number of fit parameters in the model}
\item{\bold{AIC:} AIC of the model fit}
\item{\bold{AICc:} AICc of the model fit}
\item{\bold{BIC:} BIC of the model fit}
\item{and some other potentially useful things}
}
}
\description{
Non-linear fits of different models to the element trajectories
}
\details{
the model likelihood is maximized using methods available in \code{\link[stats]{optim}}.  Optimization methods to be used within optim can be specified through the control object (i.e., control$method).  The default method is L-BFGS-B with bounds specific to each model. Each model
}
\examples{
data(pineneedles)
fit <- fit_element(mass.remaining=pineneedles$Mass.remaining,
element.remaining=c(1.1,1.2,1.3,1,0.8,0.6),
model="import.model",iters=1000)
}
\seealso{
\code{\link{optim}},
\code{\link{fit_litter}},
\code{\link{steady_state}},
\code{\link{plot.litfit}}
}
\author{
Will Cornwell
}
