\docType{package}
\name{litterfitter}
\alias{litterfitter}
\alias{litterfitter-litterfitter}
\alias{litterfitter-package}
\title{litterfitter: fit models of decomposition to single cohort litter data}
\description{
  

  litterfitter fits different models of decomposition to single cohort mass loss data.
}
\references{
  Cornwell, W. K., & Weedon, J. T. (2013). Decomposition trajectories of diverse litter types: a model selection analysis. Methods in Ecology and Evolution.
  
  Manzoni, S., Pineiro, G., Jackson, R. B., Jobbagy, E. G., Kim, J. H., & Porporato, A. (2012). Analytical models of soil and litter decomposition: Solutions for mass loss and time-dependent decay rates. Soil Biology and Biochemistry, 50, 66-76.
}

