% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/litterfitter-package.R
\docType{package}
\name{litterfitter}
\alias{litterfitter}
\alias{litterfitter-litterfitter}
\alias{litterfitter-package}
\title{litterfitter: Evaluate model adequacy for phylogenetic models of continuous character evolution}
\description{
Fitting decomposition curves to single cohort litter studies
}
\references{
Cornwell, W. K., and J. T. Weedon. (2013). Decomposition trajectories of diverse litter types: a model selection analysis. Methods in Ecology and Evolution.

Manzoni, S., Pineiro, G., Jackson, R. B., Jobbagy, E. G., Kim, J. H., & Porporato, A. (2012). Analytical models of soil and litter decomposition: Solutions for mass loss and time-dependent decay rates. Soil Biology and Biochemistry, 50, 66-76.
}

