% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pineneedles-data.R
\docType{data}
\name{pineneedles}
\alias{pineneedles}
\title{decomposition trajectory for pine needles}
\format{
a data.frame with two columns
}
\source{
Hobbie et al.
}
\usage{
data(pineneedles)
}
\description{
data from Hobbie et al means of pine needle decomposition
}
\author{
Will Cornwell
}
