% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bootstrap.parameters.R
\name{plot.litfit_bootstrap}
\alias{plot.litfit_bootstrap}
\title{Plot}
\usage{
\method{plot}{litfit_bootstrap}(x,parameter,...)
}
\arguments{
\item{x}{litfit object}

\item{...}{additional arguments passed to plot.default}

\item{parameter}{which coeficient to plot from the litfit object}
}
\description{
Plot a bootstrap distribution of a particular coefficient
}
\details{
Some more details
}
\examples{
fit <- fit_litter(time=pineneedles$Year, mass.remaining=pineneedles$Mass.remaining, model='neg.exp', iters=1000)
boot1 <- bootstrap.parameters(fit, nboot = 5000)
plot(boot1)
}
\author{
James Weedon
}
\seealso{
\code{\link{fit_litter}} \code{\link{bootstrap.parameters}}
}

