% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output-fxns.R
\name{plot_95CI}
\alias{plot_95CI}
\title{Plot model fits with 95% CI intervals}
\usage{
plot_95CI(x,repetition,obs.time)
}
\arguments{
\item{x}{litfit object}

\item{repetition}{number of repetition for bootstrap}

\item{obs.time}{maximum observation time}
}
\description{
Plot model fits with 95% CI intervals
}
\details{
this function is designed to plot 95% CI intervals from bootstrap of the model fit
}
\examples{
data(pineneedles,package='litterfitter')

fit<-fit_litter(time=pineneedles$Year,mass.remaining=pineneedles$Mass.remaining,
 model='weibull',iters=1000)
plot_95CI(fit,1000,6)

}
\seealso{
\code{\link{fit_litter}} \code{\link{bootstrap_parameters}}
}
\author{
Baptiste Wijas
}
