% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{predict_from_fit}
\alias{predict_from_fit}
\title{Predict Using a Specific Model Function}
\usage{
predict_from_fit(model_fn, params, target_time)
}
\arguments{
\item{model_fn}{A function representing the model used for prediction.}

\item{params}{A list of parameters to be passed to the model function.}

\item{target_time}{A numeric value representing the target time for the prediction.}
}
\value{
Numeric. The predicted value computed by the model function.
}
\description{
This function computes predictions by applying the specified model function (\code{model_fn})
with provided parameters (\code{params}) at a target time (\code{target_time}). It is an internal
utility function and is NOT a generic or S3 method.
}
\keyword{internal}
