% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/output-fxns.R
\name{steady_state.litfit}
\alias{steady_state.litfit}
\title{Steady-state biomass}
\usage{
\method{steady_state}{litFit}(x,...)
}
\arguments{
\item{x}{litfit object}

\item{...}{additional parameters (remove?)}
}
\description{
Estimate the steady state biomass as a proportion of the annual input,
based on the particular model fit.
}
\details{
Right now only implemented for a subset of models.  More coming soon...
}
\examples{
data(pineneedles)
fit<-fit_litter(time=pineneedles$Year,mass.remaining=pineneedles$Mass.remaining,model="neg.exp",iters=1000)
steady_state(fit)
}
\author{
Will Cornwell
}
\seealso{
\code{\link{fit_litter}}
}

