\name{steadyState.litFit}
\alias{steadyState}
\title{Steady state calculation for a particular fit to data}
\usage{
 \method{steadyState}{litFit}(x,...)
}
\arguments{
  	\item{x}{litFit object}
  	\item{...}{additional arguments}
}
\value{
  test
}
\description{
  Testing
}
\details{
  This function is an example of a help file
}
\examples{
fit<-litterFitter(time=c(0,1,2,3,4,5,6),mass.remaining=c(1,0.9,1.01,0.4,0.6,0.2,0.01),"neg.exp",iters=250)
}

