% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/output-fxns.R
\name{steady_state}
\alias{steady_state}
\title{Steady-state estimating from a lit fit object}
\usage{
steady_state(x,...)
}
\arguments{
\item{x}{litfit object}

\item{...}{additional parameters}
}
\description{
Estimate the steady state biomass as a proportion of the annual input,
based on the particular model fit.
}
\details{
Right now only implemented for a subset of models.  More coming soon...
}
\examples{
fit<-fit_litter(time=c(0,1,2,3,4,5,6),mass.remaining=c(1,0.9,1.01,0.4,0.6,0.2,0.01),"neg.exp",iters=250)
steady_state(fit)
}
\author{
Will Cornwell
}
\seealso{
\code{\link{fit_litter}}
}

